requirements.txt
azure-functions
azure-storage-blob
azure-keyvault-secrets
azure-identity
databricks-sql-connector
pandas
paramiko
python-dotenv
pysftp
pytz